removed some uneeded files/code and found out how to re-implement the dimensions

+++ src/main/java/com/Ultra_Nerd/CodeLyokoRemake15/blocks/DigitalLavaBlock.java
@@ -29,7 +29,7 @@ public class DigitalLavaBlock extends LiquidBlock {
     public void entityInside(@Nonnull BlockState state, @Nonnull Level worldIn, @Nonnull BlockPos pos, @Nonnull Entity entityIn) {
         super.entityInside(state, worldIn, pos, entityIn);
         entityIn.setRemainingFireTicks(100);
-        entityIn.hurt(new DamageSource(Objects.requireNonNull(this.getRegistryName()).toString()).bypassArmor().setIsFire(),Integer.MAX_VALUE);
+        entityIn.hurt(new DamageSource(Objects.requireNonNull(this.getRegistryName()).toString()).bypassArmor().setIsFire(),Float.MAX_VALUE);
     }
 
     @Override
